backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
numpy==1.24.3
faiss-cpu==1.7.4
pydantic==2.5.0